+++ src/common/services/recordGenerator/processors/profileSchema/formatters/value/dropdownValueFormatter.ts
@@ -4,9 +4,9 @@ import { BaseValueFormatter } from './baseValueFormatter';
 
 export class DropdownValueFormatter extends BaseValueFormatter {
   formatSimple(value: UserValueContents) {
-    return {
-      [BFLITE_URIS.LINK]: [value.meta?.uri ?? ''],
-      [BFLITE_URIS.LABEL]: [value.meta?.basicLabel ?? value.label ?? ''],
-    };
+    const uri = value.meta?.uri;
+    const label = value.meta?.basicLabel ?? value.label ?? '';
+
+    return this.buildLinkLabelObject(BFLITE_URIS.LABEL, label, uri);
   }
 }