+++ src/common/services/recordGenerator/processors/profileSchema/formatters/value/groupValueFormatter.ts
@@ -4,23 +4,23 @@ import { BaseValueFormatter } from './baseValueFormatter';
 
 export class GroupValueFormatter extends BaseValueFormatter {
   formatSimple(value: UserValueContents, recordSchemaEntry?: RecordSchemaEntry) {
-    if (!value.meta?.uri) return [];
+    const uri = value.meta?.uri;
+    const label = value.meta?.basicLabel ?? value.label ?? '';
 
     if (recordSchemaEntry?.options?.mappedValues) {
       const mappedUri = Object.entries(recordSchemaEntry.options.mappedValues).find(
-        ([, mappedValue]) => mappedValue.uri === value.meta?.uri,
+        ([, mappedValue]) => mappedValue.uri === uri,
       )?.[0];
 
       if (mappedUri) return [mappedUri];
     }
 
     if (recordSchemaEntry?.options?.includeTerm) {
-      return {
-        [BFLITE_URIS.LINK]: [value.meta.uri],
-        [BFLITE_URIS.TERM]: [value.meta?.basicLabel ?? value.label ?? ''],
-      };
+      return this.buildLinkLabelObject(BFLITE_URIS.TERM, label, uri);
     }
 
-    return [value.meta.uri];
+    if (uri) return [uri];
+
+    return [];
   }
 }